+++ client/package.json
@@ -27,6 +27,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.56.4",
+    "react-toastify": "^11.0.5",
     "recharts": "^2.15.3",
     "shadcn": "^2.5.0",
     "sonner": "^2.0.3",
